Code review comments - extract to many classes, many test files.
