+++ bin/run-test
@@ -1,4 +1,4 @@
 #! /bin/bash
 
 # allocate a tty for better test output even though not strictly needed.
-docker run -it -v /var/run/docker.sock:/var/run/docker.sock -v $PWD/universal-application-tool-0.0.1:/usr/src/universal-application-tool-0.0.1 uat sbt test
+docker run -it --rm -v /var/run/docker.sock:/var/run/docker.sock -v $PWD/universal-application-tool-0.0.1:/usr/src/universal-application-tool-0.0.1 uat sbt test
