+++ universal-application-tool-0.0.1/app/models/Applicant.java
@@ -1,6 +1,8 @@
 package models;
 
 import io.ebean.annotation.DbJsonB;
+import io.ebean.text.json.EJson;
+import java.io.IOException;
 import java.util.Map;
 import javax.persistence.Entity;
 import play.data.validation.Constraints;
@@ -8,15 +10,18 @@ import play.data.validation.Constraints;
 @Entity
 /** The ebeans mapped class that represents an individual applicant */
 public class Applicant extends BaseModel {
-  private static final long serialVersionUID = 1;
+  private static final long serialVersionUID = 1L;
 
   @Constraints.Required @DbJsonB
   // When we build an object that Jackson can deserialize, we replace Map<String, Object> with that
   // type.
   // For now, this will be automatically deserialized - with subobjects being "Map<String, Object>"
-  // and lists
-  // being List<Object>.
+  // and lists being List<Object>.
   public Map<String, Object> object;
 
   // Play will autogenerate getters and setters, unless it detects that any have been written.
+
+  public String objectAsJsonString() throws IOException {
+    return EJson.write(object);
+  }
 }
