+++ universal-application-tool-0.0.1/app/models/Question.java
@@ -1,6 +1,8 @@
 package models;
 
 import io.ebean.annotation.DbJsonB;
+import io.ebean.text.json.EJson;
+import java.io.IOException;
 import java.util.Map;
 import javax.persistence.Entity;
 import play.data.validation.Constraints;
@@ -13,9 +15,12 @@ public class Question extends BaseModel {
   // When we build an object that Jackson can deserialize, we replace Map<String, Object> with that
   // type.
   // For now, this will be automatically deserialized - with subobjects being "Map<String, Object>"
-  // and lists
-  // being List<Object>.
+  // and lists being List<Object>.
   public Map<String, Object> object;
 
   // Play will autogenerate getters and setters, unless it detects that any have been written.
+
+  public String objectAsJsonString() throws IOException {
+    return EJson.write(object);
+  }
 }
