+++ universal-application-tool-0.0.1/app/repository/QuestionRepository.java
@@ -1,0 +1,61 @@
+package repository;
+
+import static java.util.concurrent.CompletableFuture.supplyAsync;
+
+import io.ebean.Ebean;
+import io.ebean.EbeanServer;
+import java.util.Optional;
+import java.util.Set;
+import java.util.concurrent.CompletionStage;
+import javax.inject.Inject;
+import models.Applicant;
+import models.Program;
+import models.Question;
+import play.db.ebean.EbeanConfig;
+
+public class QuestionRepository {
+
+  private final EbeanServer ebeanServer;
+  private final DatabaseExecutionContext executionContext;
+
+  @Inject
+  public QuestionRepository(EbeanConfig ebeanConfig, DatabaseExecutionContext executionContext) {
+    this.ebeanServer = Ebean.getServer(ebeanConfig.defaultServer());
+    this.executionContext = executionContext;
+  }
+
+  public CompletionStage<Set<Question>> listQuestions() {
+    return supplyAsync(() -> ebeanServer.find(Question.class).findSet(), executionContext);
+  }
+
+  public CompletionStage<Optional<Question>> lookupQuestion(long id) {
+    return supplyAsync(
+        () -> Optional.ofNullable(ebeanServer.find(Question.class).setId(id).findOne()),
+        executionContext);
+  }
+
+  public CompletionStage<Optional<Question>> lookupQuestion(String target) {
+    return supplyAsync(
+        () ->
+            Optional.ofNullable(
+                ebeanServer
+                    .find(Question.class)
+                    .where()
+                    .jsonEqualTo("object", "target", target)
+                    .findOne()),
+        executionContext);
+  }
+
+  public CompletionStage<Void> insertQuestion(Question question) {
+    return supplyAsync(
+        () -> {
+          ebeanServer.insert(question);
+          return null;
+        },
+        executionContext);
+  }
+
+  public void insertQuestionSync(Question question) {
+    ebeanServer.insert(question);
+  }
+}
