+++ universal-application-tool-0.0.1/conf/evolutions/default/3.sql
@@ -10,21 +10,6 @@ create table if not exists applicants (
 create table if not exists questions (
   id bigserial primary key,
   object jsonb not null,
-  -- ensure that all entries include a 'target' key.
-  check (
-    (object->'target') is not null
-      and
-    jsonb_typeof(object->'target') = 'string'
-      and
-    LENGTH((object->>'target')::varchar) > 0
-  )
-);
-
-create unique index if not exists question_target on questions(
-  -- Create an index on 'target' - ensure that no two entries have
-  -- the same 'target' value.  This will allow us to look up what
-  -- question can help us gather a piece of necessary information.
-  (object ->>'target')
 );
 
 create table if not exists programs (
